Remove EDAs' configuration files in the config directory
We removed the EDAs' configuration files that were in the "config" directory
at the top level in the EOP code. This directory contained the same files that are
present in core/src/main/resources/configuration-file/ but with the path of the used resources
(e.g. WordNet) and models updated (by an automatic procedure) to work with the files that we distribute as part of the
eop-resources archive. In the past we thought that the files in this directory could be useful
but now we think that we can take them from core/src/main/resources/configuration-file/ directly
when we produce the eop-resources archive.