Change the mechanism to manage the created models.
This update has been requested by GN and this is his description:

Our current strategy  is that if a model file XXX specified  in the config file with same name exists, we first copy the old XXX to XXX_old.
If XXX_old already exists, we delete it, and then rename XXX to XXX_old, so that we always keep only one XXX_old.
The latter just to be on the save side and save at least the latest existing model file of same name as that used in the  config file.