Add AdArte EDA
This is a preliminary release to the public and can/will have minor problems. Please be patient and report
any bug you find with all the detail you can.

AdArte is based on modeling the Entailment Relations (i.e., Entailment, Not-Entailment) as a
classification problem. First texts (T) are mapped into hypothesis (H) by sequences of editing operations
(i.e., insertion, deletion, substitution of text portions) needed to transform T into H, where each edit
operation has a cost associated with it. Then, and this is different from the algorithms which use these
operations to calculate a threshold value that best separates the Entailment Relations from the Not-Entailment
ones, the proposed algorithm uses the calculated operations as a feature set to feed a Supervised Learning
Classifier System being able to classify the relations between T and H.