Update the LAP pom.xml file
Added a dependency to the library de.tudarmstadt.ukp.dkpro.core.api.lexmorph-asl distributed by FBK containing the correct Italian part-of-speech mapping (it-stein-tagger.map file). The wrong one contained in DKPro 1.1.4 maps all the part-of-speech with POS while the correct one distributed with the EOP (under the lap resources directory) is regularly overwritten by the DKPro one. When we will adopt DKPro 1.1.7 this library will not be necessary any more and will be removed.